// src/types/constructionSite.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface ConstructionSiteDTO {
	id: number;
	name: string;
	is_active: boolean;
}

export interface ConstructionSite {
	id: number;
	name: string;
	is_active: boolean;
}

export type ConstructionSiteKey = Pick<
	ConstructionSite,
	"id"
>;
export type ConstructionSiteNew = Pick<
	ConstructionSite,
	"name" | "is_active"
>;
export type ConstructionSiteUpd = Partial<Pick<
	ConstructionSite,
	"name" | "is_active"
>>;

export interface ConstructionSiteUpdate {
	key: ConstructionSiteKey;
	model: ConstructionSiteUpd;
}