// src/types/material.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialDTO {
	id: number;
	name: string;
	name_full: string | null;
	measure_unit_id: number;
	material_type_id: number;
	is_active: boolean;
}

export interface Material {
	id: number;
	name: string;
	name_full: string | null;
	measure_unit_id: number;
	material_type_id: number;
	is_active: boolean;
}

export type MaterialKey = Pick<
	Material,
	"id"
>;
export type MaterialNew = Pick<
	Material,
	"name" | "name_full" | "measure_unit_id" | "material_type_id" | "is_active"
>;
export type MaterialUpd = Partial<Pick<
	Material,
	"name" | "name_full" | "measure_unit_id" | "material_type_id" | "is_active"
>>;

export interface MaterialUpdate {
	key: MaterialKey;
	model: MaterialUpd;
}