// src/types/materialType.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialTypeDTO {
	id: number;
	code: string;
	name: string;
	is_active: boolean;
}

export interface MaterialType {
	id: number;
	code: string;
	name: string;
	is_active: boolean;
}

export type MaterialTypeKey = Pick<
	MaterialType,
	"id"
>;
export type MaterialTypeNew = Pick<
	MaterialType,
	"code" | "name" | "is_active"
>;
export type MaterialTypeUpd = Partial<Pick<
	MaterialType,
	"code" | "name" | "is_active"
>>;

export interface MaterialTypeUpdate {
	key: MaterialTypeKey;
	model: MaterialTypeUpd;
}