// src/types/measureUnit.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MeasureUnitDTO {
	id: number;
	name: string;
	name_full: string;
	okei_code: string;
	is_active: boolean;
}

export interface MeasureUnit {
	id: number;
	name: string;
	name_full: string;
	okei_code: string;
	is_active: boolean;
}

export type MeasureUnitKey = Pick<
	MeasureUnit,
	"id"
>;
export type MeasureUnitNew = Pick<
	MeasureUnit,
	"name" | "name_full" | "okei_code" | "is_active"
>;
export type MeasureUnitUpd = Partial<Pick<
	MeasureUnit,
	"name" | "name_full" | "okei_code" | "is_active"
>>;

export interface MeasureUnitUpdate {
	key: MeasureUnitKey;
	model: MeasureUnitUpd;
}