// src/types/supplier.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface SupplierDTO {
	id: number;
	name: string;
	name_full: string;
	inn: string;
	kpp: string | null;
	is_active: boolean;
}

export interface Supplier {
	id: number;
	name: string;
	name_full: string;
	inn: string;
	kpp: string | null;
	is_active: boolean;
}

export type SupplierKey = Pick<
	Supplier,
	"id"
>;
export type SupplierNew = Pick<
	Supplier,
	"name" | "name_full" | "inn" | "kpp" | "is_active"
>;
export type SupplierUpd = Partial<Pick<
	Supplier,
	"name" | "name_full" | "inn" | "kpp" | "is_active"
>>;

export interface SupplierUpdate {
	key: SupplierKey;
	model: SupplierUpd;
}